Major update to attendance and student course evaluation
Replace grade_preference with course_preference and add new option for
"exempt" so students who are not even taking a course can have the
value. Also allows attendance to include or exclude exempt students so
when exempt students are taken out of a class, their attendance can be
recorded by a special needs instructor.